requirements.txt

certifi==2022.12.7
charset-normalizer==3.1.0
filelock==3.10.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sklearn==0.0.post1
sympy==1.11.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
